Fixed wavetable widths and adjusted synth_audio appropriately. Builds, not tested

+++ main/waveform_tables.h
@@ -1,10 +1,10 @@
 #ifndef WAVEFORM_TABLES_H_
 #define WAVEFORM_TABLES_H_
 
-extern uint_fast16_t sin_array[256];
-extern uint_fast16_t sawtooth_array[256];
-extern uint_fast16_t square_array[256];
-extern uint_fast16_t triangle_array[256];
+extern uint_fast8_t sin_array[256];
+extern uint_fast8_t sawtooth_array[256];
+extern uint_fast8_t square_array[256];
+extern uint_fast8_t triangle_array[256];
 extern uint32_t ratio_num[128];
 extern uint32_t ratio_denom[128];
